+++ packages/@expo/cli/src/export/embed/exportEmbedAsync.ts
@@ -127,7 +127,7 @@ export async function createMetroServerAndBundleRequestAsync(
     ...Server.DEFAULT_BUNDLE_OPTIONS,
     ...getMetroDirectBundleOptionsForExpoConfig(projectRoot, exp, {
       splitChunks: false,
-      mainModuleName: options.entryFile,
+      mainModuleName: resolveRealEntryFilePath(projectRoot, options.entryFile),
       platform: options.platform,
       minify: options.minify,
       mode: options.dev ? 'development' : 'production',
@@ -227,4 +227,18 @@ export async function exportEmbedAssetsAsync(
 
 function isError(error: any): error is Error {
   return error instanceof Error;
+}
+
+/**
+ * This is a workaround for Metro not resolving entry file paths to their real location.
+ * When running exports through `eas build --local` on macOS, the `/var/folders` path is used instead of `/private/var/folders`.
+ *
+ * See: https://github.com/expo/expo/issues/28890
+ */
+function resolveRealEntryFilePath(projectRoot: string, entryFile: string): string {
+  if (projectRoot.startsWith('/private/var') && entryFile.startsWith('/var')) {
+    return fs.realpathSync(entryFile);
+  }
+
+  return entryFile;
 }