+++ packages/@expo/fingerprint/src/Options.ts
@@ -2,7 +2,9 @@ import fs from 'fs/promises';
 import os from 'os';
 import path from 'path';
 
+import { loadConfigAsync } from './Config';
 import type { NormalizedOptions, Options } from './Fingerprint.types';
+import { SourceSkips } from './sourcer/SourceSkips';
 
 export const FINGERPRINT_IGNORE_FILENAME = '.fingerprintignore';
 
@@ -78,12 +80,18 @@ export async function normalizeOptionsAsync(
   projectRoot: string,
   options?: Options
 ): Promise<NormalizedOptions> {
+  const config = await loadConfigAsync(projectRoot, options?.silent ?? false);
   return {
-    ...options,
-    platforms: options?.platforms ?? ['android', 'ios'],
-    concurrentIoLimit: options?.concurrentIoLimit ?? os.cpus().length,
-    hashAlgorithm: options?.hashAlgorithm ?? 'sha1',
+    // Defaults
+    platforms: ['android', 'ios'],
+    concurrentIoLimit: os.cpus().length,
+    hashAlgorithm: 'sha1',
     ignorePaths: await collectIgnorePathsAsync(projectRoot, options),
+    sourceSkips: SourceSkips.None,
+    // Options from config
+    ...config,
+    // Explicit options
+    ...options,
   };
 }