+++ packages/expo-sensors/src/__tests__/Accelerometer-test.ios.ts
@@ -8,7 +8,7 @@ it(`notifies listeners`, () => {
   const mockListener = jest.fn();
   Accelerometer.addListener(mockListener);
 
-  const mockEvent = { x: 0.2, y: 0.1, z: 0.3 };
+  const mockEvent = { x: 0.2, y: 0.1, z: 0.3, timestamp: 123456 };
   Accelerometer._nativeModule.emit('accelerometerDidUpdate', mockEvent);
   expect(mockListener).toHaveBeenCalledWith(mockEvent);
 });